requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage
django-debug-toolbar==1.3.2
